refactor: migrate webhome home page to app
