+++ apps/web/public/images/hero-bg.jpg
