+++ apps/web/src/components/layout/layout.tsx
@@ -1,15 +1,16 @@
+import { Footer } from './footer';
+import { Header } from './header';
 import { useUser } from '@stackframe/react';
 import { useLocation } from '@tanstack/react-router';
 import { useEffect } from 'react';
 import { sendIdentify } from '~/external/segment';
 import { isAppPage } from '~/utils/router-checker';
-import { Footer } from './footer';
-import { Header } from './header';
 
 export function Layout({ children }: { children: React.ReactNode }) {
   const user = useUser();
   const { pathname } = useLocation();
   const hideFooter = isAppPage(pathname);
+  const isPublicHome = pathname === '/' && !user;
 
   useEffect(() => {
     if (user?.id) {
@@ -17,11 +18,21 @@ export function Layout({ children }: { children: React.ReactNode }) {
     }
   }, [user?.id]);
 
+  const content = (
+    <>
+      <Header />
+      <main className="flex-1 flex flex-col">{children}</main>
+      <Footer isHidden={hideFooter} />
+    </>
+  );
+
+  if (isPublicHome) {
+    return content;
+  }
+
   return (
     <div className="mx-auto flex flex-col h-screen w-5/6 md:w-4/5 gap-2 overflow-hidden">
-      <Header />
-      <main className="h-screen overflow-y-auto">{children}</main>
-      <Footer isHidden={hideFooter} />
+      {content}
     </div>
   );
 }
