--- apps/web/src/components/layout/public/public-layout.tsx
@@ -1,16 +1,0 @@
-import { Footer } from '../footer';
-import { PublicHeader } from './public-header';
-
-interface PublicLayoutProps {
-  children: React.ReactNode;
-}
-
-export function PublicLayout({ children }: PublicLayoutProps) {
-  return (
-    <>
-      <PublicHeader />
-      <main className="flex-1 flex flex-col">{children}</main>
-      <Footer />
-    </>
-  );
-}
