+++ apps/web/src/pages/app-page.tsx
@@ -18,7 +18,7 @@ export function AppPage() {
   const { isLoading } = useApp(appId);
 
   useEffect(() => {
-    sendPageView(AnalyticsEvents.PAGE_VIEW_CHAT);
+    sendPageView(AnalyticsEvents.PAGE_VIEW_APP);
   }, []);
 
   const renderContent = () => {
@@ -33,7 +33,7 @@ export function AppPage() {
   };
 
   return (
-    <div className="flex flex-col items-center w-full h-full">
+    <div className="flex flex-col items-center w-full h-full mt-24">
       {renderContent()}
       <div
         className="fixed bottom-5 left-1/2 transform -translate-x-1/2 w-4/5 max-w-4xl"
