+++ apps/web/src/pages/home/home-page.tsx
@@ -1,5 +1,5 @@
-import { AuthenticatedHome } from './home/authenticated-home';
-import { PublicHome } from './home/public-home';
+import { AuthenticatedHome } from './authenticated-home';
+import { PublicHome } from './public-home';
 import { useUser } from '@stackframe/react';
 import { createLazyRoute } from '@tanstack/react-router';
 
@@ -9,9 +9,6 @@ export const HomePageRoute = createLazyRoute('/')({
 
 export function HomePage() {
   const user = useUser();
-  if (user) {
-    return <AuthenticatedHome />;
-  }
 
-  return <PublicHome />;
+  return user ? <AuthenticatedHome /> : <PublicHome />;
 }
